+++ src/Pencil/Blog.hs
@@ -8,6 +8,7 @@ module Pencil.Blog where
 import Pencil.Config
 import Pencil.Env
 import Pencil.App
+import Pencil.Content
 import Pencil.Internal.Pencil
 import Pencil.Parser.Internal as Parser
 import Control.Monad (liftM, foldM)