+++ src/Pencil/Content.hs
@@ -1,0 +1,5 @@
+module Pencil.Content
+  ( module Pencil.Content.Internal
+  ) where
+
+import Pencil.Content.Internal