+++ src/Pencil/Env.hs
@@ -1,5 +1,10 @@
 module Pencil.Env
   ( module Pencil.Env.Internal
+  , merge
+  , adjust
+  , insertText
+  , insertEnv
+
   , findEnv
   , aesonToEnv
   , maybeInsertIntoEnv
@@ -7,6 +12,8 @@ module Pencil.Env
 
 import Pencil.Env.Internal
 import Pencil.Parser.Internal
+import Pencil.Content.Internal
+import Pencil.App.Internal
 
 import Data.Text.Encoding (encodeUtf8)
 
@@ -14,6 +21,43 @@ import qualified Data.Maybe as M
 import qualified Data.HashMap.Strict as H
 import qualified Data.Yaml as A
 import qualified Data.Text as T
+
+-- | Merges two @Env@s together, biased towards the left-hand @Env@ on duplicates.
+merge :: Env -> Env -> Env
+merge = H.union
+
+-- | Inserts text into the given @Env@.
+--
+-- @
+-- env <- asks getEnv
+-- insertText "title" "My Awesome Website" env
+-- @
+insertText :: T.Text
+           -- ^ Environment variable name.
+           -> T.Text
+           -- ^ Text to insert.
+           -> Env
+           -- ^ Environment to modify.
+           -> Env
+insertText var val = H.insert var (VText val)
+
+-- | Modifies a variable in the given environment.
+adjust :: (Value -> Value)
+       -> T.Text
+       -- ^ Environment variable name.
+       -> Env
+       -> Env
+adjust = H.adjust
+
+-- | Inserts @Value@ into the given @Env@.
+insertEnv :: T.Text
+          -- ^ Environment variable name.
+          -> Value
+          -- ^ @Value@ to insert.
+          -> Env
+          -- ^ Environment to modify.
+          -> Env
+insertEnv = H.insert
 
 -- | Find preamble node, and load as an Env. If no preamble is found, return a
 -- blank Env.