+++ test/Spec.hs
@@ -2,4 +2,4 @@ import Test.DocTest
 
 main :: IO ()
 main =
-  doctest ["-isrc", "src/Pencil/Internal/Parser.hs"]
+  doctest ["-isrc", "src/Pencil/Parser/Internal.hs"]